backend/app/pyproject.toml
[tool.poetry]
name = "fastapi_graphql"
version = "0.1.0"
description = "testing fastapi with graphql"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
fastapi = "^0.78.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.116.3"}
SQLAlchemy = "^1.4.39"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
